requirements.txt
fastapi
uvicorn
pyyaml
requests
openai
langchain
langchain-google-genai
python-dotenv
ics
graph
langchain_groq